Fix ReadBank not returning number of bytes read
The ReadBank method incorrectly incremented the pointer
value of num_bytes_read instead of the value, so the
method would never modify the actual value. This leads
to the SaveFile method being called with a length of 0
bytes, which is not what we want. :)
VerifyBank was affected in a similar way.